Replace commit author name with login (#175)
GitHub markdown converts @-references with links to user profiles. Using
the full author name means potentially including a space which breaks this
linking behavior. Instead of the author name (Ex: Mario Tacke), this fix
uses the author login (Ex: mariotacke) to form a proper link in the
release notes and changelog).

This fixes #158